+++ api-service/src/controllers/QueryWrapper/SqlQueryWrapper.ts
@@ -5,61 +5,15 @@ import logger from "../../logger";
 import { ResponseHandler } from "../../helpers/ResponseHandler";
 import { ErrorObject } from "../../types/ResponseModel";
 import { druidHttpService } from "../../connections/druidConnection";
-import { getDatasourceList } from "../../services/DatasourceService";
+import { getDatasourceList, getLiveDatasourcesByNames } from "../../services/DatasourceService";
 import { AxiosResponse } from "axios";
 import { Parser } from "node-sql-parser";
-import { Op } from "sequelize";
-import { Datasource } from "../../models/Datasource";
+import { collectCteNames, collectTableEntries } from "../../services/SqlQueryService";
 
 const apiId = "api.obsrv.data.sql-query";
 const errorCode = "SQL_QUERY_FAILURE"
-export const result_data = {"data": {}};
+export const result_data = { "data": {} };
 const parser = new Parser();
-
-// Collect every CTE-declared name anywhere in the AST. These are local aliases
-// (WITH <name> AS ...), not datasources, so they must never be treated as tables.
-const collectCteNames = (node: any, acc: Set<string> = new Set()): Set<string> => {
-    if (_.isArray(node)) {
-        node.forEach((n) => collectCteNames(n, acc));
-        return acc;
-    }
-    if (!_.isObject(node)) return acc;
-    const withClause: any = _.get(node, "with");
-    if (_.isArray(withClause)) {
-        withClause.forEach((cte: any) => {
-            const name = _.get(cte, "name.value") || _.get(cte, "name");
-            if (_.isString(name)) acc.add(name);
-        });
-    }
-    _.forEach(node, (value) => collectCteNames(value, acc));
-    return acc;
-};
-
-// Recursively collect every FROM entry that references a real table name across
-// the whole AST (nested selects, set-ops, CTE bodies), skipping CTE-declared names.
-const collectTableEntries = (node: any, cteNames: Set<string>, acc: any[] = []): any[] => {
-    if (_.isArray(node)) {
-        node.forEach((n) => collectTableEntries(n, cteNames, acc));
-        return acc;
-    }
-    if (!_.isObject(node)) return acc;
-    const fromArr = _.get(node, "from");
-    if (_.isArray(fromArr)) {
-        _.forEach(fromArr, (entry: any) => {
-            const subquery = _.get(entry, "expr.ast");
-            if (subquery) {
-                collectTableEntries(subquery, cteNames, acc);
-                return;
-            }
-            const table = _.get(entry, "table");
-            if (_.isString(table) && !cteNames.has(table)) acc.push(entry);
-        });
-    }
-    _.forEach(node, (value, key) => {
-        if (key !== "from") collectTableEntries(value, cteNames, acc);
-    });
-    return acc;
-};
 
 // Rewrites table names to datasource_refs ONLY when a name matches a datasource
 // alias that is not itself already a datasource_ref. Any other case (name already
@@ -89,18 +43,8 @@ const resolveDatasourceQuery = async (query: string, resmsgid?: string): Promise
             return query;
         }
 
-        // Single indexed lookup for all referenced names at once (alias OR ref).
-        const rows: any[] = await Datasource.findAll({
-            where: {
-                status: "Live",
-                [Op.or]: [
-                    { datasource_ref: { [Op.in]: tableNames } },
-                    { datasource: { [Op.in]: tableNames } },
-                ],
-            },
-            attributes: ["datasource", "datasource_ref"],
-            raw: true,
-        });
+        // Only hit Postgres once we know the query actually references tables.
+        const rows = await getLiveDatasourcesByNames(tableNames);
         if (_.isEmpty(rows)) {
             logger.warn({ apiId, resmsgid, tableNames, message: "No matching datasource/datasource_ref exists, passing query as-is" });
             return query;
@@ -198,14 +142,14 @@ const fetchDruidDataSources = async (): Promise<{ TABLE_NAME: string }[]> => {
 
 const isTableSchemaQuery = (sqlQuery?: string): boolean => {
     return (
-      sqlQuery
-        ?.trim()
-        .replace(/\s+/g, " ")
-        .toUpperCase() ===
-      "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'DRUID'"
+        sqlQuery
+            ?.trim()
+            .replace(/\s+/g, " ")
+            .toUpperCase() ===
+        "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'DRUID'"
     );
-  };
-  
+};
+
 
 const createMockAxiosResponse = (data: any): AxiosResponse => {
     return {
